pyproject.toml
[tool.poetry]
name = "ip_reputation_indexer"
version = "0.1.0"
description = "get reputation for ip"
authors = ["msterhuj <gabin.lanore@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
PyYAML = "^6.0"
rich = "^13.3.5"
sqlmodel = "^0.0.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
